Criação do provider base para requisições HTTP e sepração dos componentes de usuários

+++ src/app/clientes/clientes.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { ClienteService } from './cliente-service';
-import { Ng4LoadingSpinnerService } from 'ng4-loading-spinner';
+// import { Ng4LoadingSpinnerService } from 'ng4-loading-spinner';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ClienteModel } from './cliente.model';
+import { NgxUiLoaderService } from 'ngx-ui-loader';
 
 
 @Component({
@@ -17,26 +18,40 @@ export class ClientesComponent implements OnInit {
 isEditable: boolean;
 @Output() public add = new EventEmitter();
 
-  constructor(private clienteSrvc:ClienteService, private route: Router) { }
+  constructor(private clienteSrvc: ClienteService,
+              private route: Router,
+              private ngxService: NgxUiLoaderService
+              ) { }
 
   ngOnInit() {
   }
-onEdit(){
+onEdit() {
   this.isEditable = true;
 }
-offEdit(){
-  this.isEditable = false
+offEdit() {
+  this.isEditable = false;
 }
-deleteClient(){
-  this.clienteSrvc.deleteClient(this.cliente._id).subscribe(()=>{
-    console.log(`Cliente excluido`)
-    this.route.navigate(['/usuarios'])
-  })
+deleteClient() {
+
+  let r = confirm(`Deseja realmente excluir o cliente ${this.cliente.nome} ?`);
+
+  if (r == true) {
+  this.clienteSrvc.deleteClient(this.cliente._id).subscribe(() => {
+    console.log(`Cliente excluido`);
+    this.ngxService.stop();
+    window.location.reload();
+    // this.route.navigate(['/usuarios']);
+  });
 }
-  emitAddEvent(){
-   this.add.emit(this.cliente)
-   this.clienteSrvc.alterClient(this.cliente, this.cliente._id).subscribe(v=>{
-     console.log(`Deu certo !!`)
-   })
+}
+  emitAddEvent() {
+   this.add.emit(this.cliente);
+   console.log('cliente', this.cliente);
+   this.clienteSrvc.alterClient(this.cliente, this.cliente._id).subscribe(v => {
+     this.isEditable = false;
+    //  this.ngxService.stop();
+     window.location.reload();
+
+    });
   }
 }
